+++ libraries/exoplayer_hls/src/main/java/androidx/media3/exoplayer/hls/HlsMediaSource.java
@@ -189,41 +189,40 @@ public final class HlsMediaSource extends BaseMediaSource
     }
 
     /**
-     * Sets the factory from which playlist parsers will be obtained. The default value is a {@link
-     * DefaultHlsPlaylistParserFactory}.
+     * Sets the factory from which playlist parsers will be obtained.
      *
      * @param playlistParserFactory An {@link HlsPlaylistParserFactory}.
      * @return This factory, for convenience.
      */
-    public Factory setPlaylistParserFactory(
-        @Nullable HlsPlaylistParserFactory playlistParserFactory) {
+    public Factory setPlaylistParserFactory(HlsPlaylistParserFactory playlistParserFactory) {
       this.playlistParserFactory =
-          playlistParserFactory != null
-              ? playlistParserFactory
-              : new DefaultHlsPlaylistParserFactory();
+          checkNotNull(
+              playlistParserFactory,
+              "HlsMediaSource.Factory#setPlaylistParserFactory no longer handles null by"
+                  + " instantiating a new DefaultHlsPlaylistParserFactory. Explicitly"
+                  + " construct and pass an instance in order to retain the old behavior.");
       return this;
     }
 
     /**
-     * Sets the {@link HlsPlaylistTracker} factory. The default value is {@link
-     * DefaultHlsPlaylistTracker#FACTORY}.
+     * Sets the {@link HlsPlaylistTracker} factory.
      *
      * @param playlistTrackerFactory A factory for {@link HlsPlaylistTracker} instances.
      * @return This factory, for convenience.
      */
-    public Factory setPlaylistTrackerFactory(
-        @Nullable HlsPlaylistTracker.Factory playlistTrackerFactory) {
+    public Factory setPlaylistTrackerFactory(HlsPlaylistTracker.Factory playlistTrackerFactory) {
       this.playlistTrackerFactory =
-          playlistTrackerFactory != null
-              ? playlistTrackerFactory
-              : DefaultHlsPlaylistTracker.FACTORY;
+          checkNotNull(
+              playlistTrackerFactory,
+              "HlsMediaSource.Factory#setPlaylistTrackerFactory no longer handles null by"
+                  + " defaulting to DefaultHlsPlaylistTracker.FACTORY. Explicitly"
+                  + " pass a reference to this instance in order to retain the old behavior.");
       return this;
     }
 
     /**
      * Sets the factory to create composite {@link SequenceableLoader}s for when this media source
-     * loads data from multiple streams (video, audio etc...). The default is an instance of {@link
-     * DefaultCompositeSequenceableLoaderFactory}.
+     * loads data from multiple streams (video, audio etc...).
      *
      * @param compositeSequenceableLoaderFactory A factory to create composite {@link
      *     SequenceableLoader}s for when this media source loads data from multiple streams (video,
@@ -231,11 +230,13 @@ public final class HlsMediaSource extends BaseMediaSource
      * @return This factory, for convenience.
      */
     public Factory setCompositeSequenceableLoaderFactory(
-        @Nullable CompositeSequenceableLoaderFactory compositeSequenceableLoaderFactory) {
+        CompositeSequenceableLoaderFactory compositeSequenceableLoaderFactory) {
       this.compositeSequenceableLoaderFactory =
-          compositeSequenceableLoaderFactory != null
-              ? compositeSequenceableLoaderFactory
-              : new DefaultCompositeSequenceableLoaderFactory();
+          checkNotNull(
+              compositeSequenceableLoaderFactory,
+              "HlsMediaSource.Factory#setCompositeSequenceableLoaderFactory no longer handles null"
+                  + " by instantiating a new DefaultCompositeSequenceableLoaderFactory. Explicitly"
+                  + " construct and pass an instance in order to retain the old behavior.");
       return this;
     }